Added updates for Shawn
Shawn asked that duplicated species be removed (we can't increase the sample size bc the source of one preprocessed the data differently). Added manual_lenght_out so that the length of the arrays for the ingroup and outgroup would be equal. (changed number of samples resulted in different number that would be evenly divisible by both sets).